# scripts/openapi-snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ForgeMsg API",
    "description": "Unified omnichannel messaging platform API",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": { "type": "apiKey", "name": "X-API-Key", "in": "header" },
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/api/v1/workflows/{id}": {
      "get": {
        "summary": "Get workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "put": {
        "summary": "Update workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Delete workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/workflows/{id}/activate": {
      "post": {
        "summary": "Activate workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/workflows/{id}/deactivate": {
      "post": {
        "summary": "Deactivate workflow",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/workflows/{id}/trigger": {
      "post": {
        "summary": "Manually trigger workflow for a contact",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/workflows/{id}/runs": {
      "get": {
        "summary": "List workflow runs",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/workflows/{id}/runs/{runId}/cancel": {
      "post": {
        "summary": "Cancel a workflow run",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true },
          { "schema": { "type": "string" }, "in": "path", "name": "runId", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
    "/api/v1/workflows/templates/{templateId}": {
      "get": {
        "summary": "Get full pre-built workflow template",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "templateId", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/workflows/templates/{templateId}/use": {
      "post": {
        "summary": "Create a workflow from a pre-built template",
        "tags": ["Workflows"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "templateId", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
    "/api/v1/holdout-groups/{id}/populate": {
      "post": {
        "tags": ["Holdout"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/holdout-groups/{id}/members": {
      "get": {
        "tags": ["Holdout"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/holdout/check/{contactId}": {
      "get": {
        "tags": ["Holdout"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "contactId", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
    "/api/v1/data-sets/{id}": {
      "get": {
        "summary": "Get a data set",
        "tags": ["Data Hub"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "patch": {
        "summary": "Update a data set",
        "tags": ["Data Hub"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "summary": "Delete a data set",
        "tags": ["Data Hub"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
    "/api/v1/data-sets/{id}/prepare": {
      "post": {
        "summary": "Bind params + return the prepared SQL (runner executes)",
        "tags": ["Data Hub"],
        "parameters": [
          { "schema": { "type": "string" }, "in": "path", "name": "id", "required": true }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    },
...
  },
  "servers": [{ "url": "http://localhost:3001", "description": "Development" }]
}